requirements.txt
django
djangorestframework
Pillow
django-crispy-forms
crispy-bootstrap5